Optimise some scripts under folder tests/ipfwd
